pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pathways"
dynamic = ["version"]
description = "Curate user pathways for The Turing Way"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "The Turing Way Community" },
]
dependencies = [
  "mdutils~=1.3",
  "markdown-it-py~=3.0",
  "PyYAML~=6.0",
]

[project.scripts]
pathways = "pathways.main:main"

[project.urls]
Documentation = "https://github.com/unknown/pathways#readme"
Issues = "https://github.com/unknown/pathways/issues"
Source = "https://github.com/unknown/pathways"

[tool.hatch.build]
only-include = ["pathways/"]

[tool.hatch.version]
path = "pathways/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
]
[tool.hatch.envs.default.scripts]
test = "python -m unittest discover {args:tests}"
test-cov = "coverage run -m unittest discover {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report --show-missing",
]
cov = [
  "test-cov",
  "cov-report",
]

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "black>=23.1.0",
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:pathways tests}"
style = [
  "ruff check {args:.}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "ruff check --fix {args:.}",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.black]
target-version = ["py311"]
skip-string-normalization = true

[tool.ruff]
target-version = "py311"
lint.select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]

[tool.ruff.lint.isort]
known-first-party = ["pathways"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["pathways", "tests"]
branch = true
parallel = true
omit = [
  "pathways/__about__.py",
]

[tool.coverage.paths]
pathways = ["pathways", "*/pathways/pathways"]
tests = ["tests", "*/pathways/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
